fix: consolidate files to satisfy pkg-file-count CI gate
Merge pause/resume handlers, service lifecycle code, and related tests into
existing package files so package file-count limits pass after the feature
additions.
