+++ pkg/factorysessionexecution/fake_service_runtime_test.go
@@ -1828,3 +1828,18 @@ func testJavaScriptRuntimeSyncCompletedEvents(t *testing.T, service *JavaScriptR
 		t.Fatalf("events = %d, want 3 canonical lifecycle events", len(events.Events))
 	}
 }
+func TestProjectedLifecycleControlStatus_PrefersCanonicalBracketStatus(t *testing.T) {
+	status := ProjectedLifecycleControlStatus("PAUSED", "RUNNING")
+	if status != LifecycleStatusPaused {
+		t.Fatalf("status = %q, want PAUSED", status)
+	}
+}
+
+func TestProjectedLifecycleControlStatus_FallsBackToFactoryRuntimeState(t *testing.T) {
+	if got := ProjectedLifecycleControlStatus("", "PAUSED"); got != LifecycleStatusPaused {
+		t.Fatalf("status = %q, want PAUSED", got)
+	}
+	if got := ProjectedLifecycleControlStatus("", "RUNNING"); got != LifecycleStatusRunning {
+		t.Fatalf("status = %q, want RUNNING", got)
+	}
+}
