--- pkg/service/live_session_lifecycle_control_test.go
@@ -1,152 +1,0 @@
-package service
-
-import (
-	"context"
-	"errors"
-	"testing"
-	"time"
-
-	factoryapi "github.com/portpowered/infinite-you/pkg/api/generated"
-	"github.com/portpowered/infinite-you/pkg/apisurface"
-	"github.com/portpowered/infinite-you/pkg/factorysessionexecution"
-	"github.com/portpowered/infinite-you/pkg/interfaces"
-)
-
-func TestFactoryService_PauseLiveFactorySession_AcceptsRunningSession(t *testing.T) {
-	harness := startRunningSessionService(t, runningSessionServiceOptions{
-		rootConfig: minimalFactoryConfig(),
-	})
-	defer harness.stop(t)
-
-	response, err := harness.svc.PauseLiveFactorySession(
-		context.Background(),
-		defaultFactorySessionID,
-		factoryapi.FactorySessionLifecycleControlRequest{},
-	)
-	if err != nil {
-		t.Fatalf("PauseLiveFactorySession: %v", err)
-	}
-	if response.Operation != factoryapi.FactorySessionLifecycleControlKindPause {
-		t.Fatalf("operation = %q, want PAUSE", response.Operation)
-	}
-	if response.Outcome != factoryapi.FactorySessionLifecycleControlOutcomeAccepted {
-		t.Fatalf("outcome = %q, want ACCEPTED", response.Outcome)
-	}
-	if response.Status != factoryapi.FactorySessionDurableLifecycleStatusPaused {
-		t.Fatalf("status = %q, want PAUSED", response.Status)
-	}
-
-	waitForSessionFactoryState(
-		t,
-		harness.svc,
-		defaultFactorySessionID,
-		interfaces.FactoryStatePaused,
-		time.Second,
-		"live session paused",
-	)
-}
-
-func TestFactoryService_ResumeLiveFactorySession_AcceptsPausedSession(t *testing.T) {
-	harness := startRunningSessionService(t, runningSessionServiceOptions{
-		rootConfig: minimalFactoryConfig(),
-	})
-	defer harness.stop(t)
-
-	if _, err := harness.svc.PauseLiveFactorySession(
-		context.Background(),
-		defaultFactorySessionID,
-		factoryapi.FactorySessionLifecycleControlRequest{},
-	); err != nil {
-		t.Fatalf("PauseLiveFactorySession: %v", err)
-	}
-
-	response, err := harness.svc.ResumeLiveFactorySession(
-		context.Background(),
-		defaultFactorySessionID,
-		factoryapi.FactorySessionLifecycleControlRequest{},
-	)
-	if err != nil {
-		t.Fatalf("ResumeLiveFactorySession: %v", err)
-	}
-	if response.Operation != factoryapi.FactorySessionLifecycleControlKindResume {
-		t.Fatalf("operation = %q, want RESUME", response.Operation)
-	}
-	if response.Outcome != factoryapi.FactorySessionLifecycleControlOutcomeAccepted {
-		t.Fatalf("outcome = %q, want ACCEPTED", response.Outcome)
-	}
-	if response.Status != factoryapi.FactorySessionDurableLifecycleStatusRunning {
-		t.Fatalf("status = %q, want RUNNING", response.Status)
-	}
-}
-
-func TestFactoryService_PauseLiveFactorySession_RepeatPauseReturnsNoOp(t *testing.T) {
-	harness := startRunningSessionService(t, runningSessionServiceOptions{
-		rootConfig: minimalFactoryConfig(),
-	})
-	defer harness.stop(t)
-
-	if _, err := harness.svc.PauseLiveFactorySession(
-		context.Background(),
-		defaultFactorySessionID,
-		factoryapi.FactorySessionLifecycleControlRequest{},
-	); err != nil {
-		t.Fatalf("PauseLiveFactorySession: %v", err)
-	}
-
-	response, err := harness.svc.PauseLiveFactorySession(
-		context.Background(),
-		defaultFactorySessionID,
-		factoryapi.FactorySessionLifecycleControlRequest{},
-	)
-	if err != nil {
-		t.Fatalf("repeat PauseLiveFactorySession: %v", err)
-	}
-	if response.Outcome != factoryapi.FactorySessionLifecycleControlOutcomeNoOp {
-		t.Fatalf("outcome = %q, want NO_OP", response.Outcome)
-	}
-	if response.Status != factoryapi.FactorySessionDurableLifecycleStatusPaused {
-		t.Fatalf("status = %q, want PAUSED", response.Status)
-	}
-}
-
-func TestFactoryService_PauseLiveFactorySession_MissingSessionReturnsNotFound(t *testing.T) {
-	harness := startRunningSessionService(t, runningSessionServiceOptions{
-		rootConfig: minimalFactoryConfig(),
-	})
-	defer harness.stop(t)
-
-	_, err := harness.svc.PauseLiveFactorySession(
-		context.Background(),
-		"live-session-missing-001",
-		factoryapi.FactorySessionLifecycleControlRequest{},
-	)
-	if err == nil {
-		t.Fatal("PauseLiveFactorySession = nil, want not found")
-	}
-	if !errors.Is(err, apisurface.ErrFactorySessionNotFound) {
-		t.Fatalf("error = %v, want ErrFactorySessionNotFound", err)
-	}
-}
-
-func TestFactoryService_ResumeLiveFactorySession_RunningSessionReturnsInvalidState(t *testing.T) {
-	harness := startRunningSessionService(t, runningSessionServiceOptions{
-		rootConfig: minimalFactoryConfig(),
-	})
-	defer harness.stop(t)
-
-	_, err := harness.svc.ResumeLiveFactorySession(
-		context.Background(),
-		defaultFactorySessionID,
-		factoryapi.FactorySessionLifecycleControlRequest{},
-	)
-	if err == nil {
-		t.Fatal("ResumeLiveFactorySession = nil, want invalid state")
-	}
-	var controlErr *factorysessionexecution.ControlError
-	if !errors.As(err, &controlErr) {
-		t.Fatalf("error = %v, want ControlError", err)
-	}
-	if controlErr.Outcome != factorysessionexecution.LifecycleControlOutcomeInvalidState {
-		t.Fatalf("outcome = %q, want INVALID_STATE", controlErr.Outcome)
-	}
-}
