+++ pkg/service/runtime_sessions.go
@@ -24,6 +24,7 @@ import (
 	"github.com/portpowered/infinite-you/pkg/factorysessions"
 	"github.com/portpowered/infinite-you/pkg/factorysessionexecution"
 	"github.com/portpowered/infinite-you/pkg/interfaces"
+	"github.com/portpowered/infinite-you/pkg/internal/metrics"
 	"github.com/portpowered/infinite-you/pkg/petri"
 	"github.com/portpowered/infinite-you/pkg/service/runtimebuild"
 	"go.uber.org/zap"
@@ -1317,4 +1318,218 @@ func (fs *FactoryService) RetryDurableFactorySessionDispatch(
 		return factoryapi.FactorySessionLifecycleControlResponse{}, err
 	}
 	return factorysession.LifecycleControlResponseToAPI(result), nil
+}
+
+func (fs *FactoryService) PauseLiveFactorySession(
+	ctx context.Context,
+	sessionID string,
+	request factoryapi.FactorySessionLifecycleControlRequest,
+) (factoryapi.FactorySessionLifecycleControlResponse, error) {
+	control, err := factorysession.ControlRequestFromAPI(request)
+	if err != nil {
+		return factoryapi.FactorySessionLifecycleControlResponse{}, err
+	}
+	result, err := fs.applyLiveLifecycleControl(
+		ctx,
+		sessionID,
+		factorysessionexecution.LifecycleControlPause,
+		control,
+	)
+	if err != nil {
+		return factoryapi.FactorySessionLifecycleControlResponse{}, err
+	}
+	return factorysession.LifecycleControlResponseToAPI(result), nil
+}
+
+func (fs *FactoryService) ResumeLiveFactorySession(
+	ctx context.Context,
+	sessionID string,
+	request factoryapi.FactorySessionLifecycleControlRequest,
+) (factoryapi.FactorySessionLifecycleControlResponse, error) {
+	control, err := factorysession.ControlRequestFromAPI(request)
+	if err != nil {
+		return factoryapi.FactorySessionLifecycleControlResponse{}, err
+	}
+	result, err := fs.applyLiveLifecycleControl(
+		ctx,
+		sessionID,
+		factorysessionexecution.LifecycleControlResume,
+		control,
+	)
+	if err != nil {
+		return factoryapi.FactorySessionLifecycleControlResponse{}, err
+	}
+	return factorysession.LifecycleControlResponseToAPI(result), nil
+}
+
+func (fs *FactoryService) applyLiveLifecycleControl(
+	ctx context.Context,
+	sessionID string,
+	operation factorysessionexecution.LifecycleControlKind,
+	control factorysessionexecution.ControlRequest,
+) (factorysessionexecution.LifecycleControlResult, error) {
+	if fs == nil {
+		return factorysessionexecution.LifecycleControlResult{}, fmt.Errorf("factory service is required")
+	}
+	if err := ctx.Err(); err != nil {
+		return factorysessionexecution.LifecycleControlResult{}, err
+	}
+
+	if _, err := factorysessionexecution.NormalizeControlRequest(control); err != nil {
+		return factorysessionexecution.LifecycleControlResult{}, err
+	}
+
+	activeFactory, err := fs.sessionFactory(sessionID)
+	if err != nil {
+		fs.observeLiveLifecycleControl(sessionID, operation, control, "", "", err)
+		return factorysessionexecution.LifecycleControlResult{}, err
+	}
+
+	snapshot, err := activeFactory.GetEngineStateSnapshot(ctx)
+	if err != nil {
+		return factorysessionexecution.LifecycleControlResult{}, fmt.Errorf("get engine state snapshot: %w", err)
+	}
+
+	currentStatus := factorysessionexecution.LifecycleStatusFromFactoryRuntimeState(snapshot.FactoryState)
+	outcome := factorysessionexecution.EvaluateLifecycleControl(operation, currentStatus)
+	if outcome == factorysessionexecution.LifecycleControlOutcomeInvalidState ||
+		outcome == factorysessionexecution.LifecycleControlOutcomeTerminalSession {
+		controlErr := &factorysessionexecution.ControlError{
+			Operation: operation,
+			Outcome:   outcome,
+			Status:    currentStatus,
+			Message: fmt.Sprintf(
+				"%s rejected for session %s in status %s",
+				operation,
+				sessionID,
+				currentStatus,
+			),
+		}
+		fs.observeLiveLifecycleControl(sessionID, operation, control, outcome, currentStatus, controlErr)
+		return factorysessionexecution.LifecycleControlResult{}, controlErr
+	}
+
+	resultStatus := currentStatus
+	if outcome == factorysessionexecution.LifecycleControlOutcomeAccepted {
+		switch operation {
+		case factorysessionexecution.LifecycleControlPause:
+			if err := activeFactory.Pause(ctx); err != nil {
+				return factorysessionexecution.LifecycleControlResult{}, fmt.Errorf("pause live factory session: %w", err)
+			}
+			resultStatus = factorysessionexecution.LifecycleStatusPaused
+		case factorysessionexecution.LifecycleControlResume:
+			if err := activeFactory.Resume(ctx); err != nil {
+				return factorysessionexecution.LifecycleControlResult{}, fmt.Errorf("resume live factory session: %w", err)
+			}
+			resultStatus = factorysessionexecution.LifecycleStatusRunning
+		default:
+			return factorysessionexecution.LifecycleControlResult{}, fmt.Errorf("unsupported live lifecycle operation %s", operation)
+		}
+	}
+
+	result := factorysessionexecution.LifecycleControlResult{
+		SessionID: sessionID,
+		Operation: operation,
+		Outcome:   outcome,
+		Status:    resultStatus,
+		Links:     factorysession.LiveLifecycleControlLinksForSession(sessionID),
+	}
+	fs.observeLiveLifecycleControl(sessionID, operation, control, outcome, resultStatus, nil)
+	return result, nil
+}
+const (
+	runtimeMetricLifecycleControl = "runtime.lifecycle_control"
+)
+
+func (fs *FactoryService) observeLiveLifecycleControl(
+	sessionID string,
+	operation factorysessionexecution.LifecycleControlKind,
+	control factorysessionexecution.ControlRequest,
+	outcome factorysessionexecution.LifecycleControlOutcome,
+	status factorysessionexecution.LifecycleStatus,
+	err error,
+) {
+	if fs == nil {
+		return
+	}
+
+	outcomeClass := lifecycleControlOutcomeClass(outcome, err)
+	fields := liveLifecycleControlLogFields(sessionID, operation, outcomeClass, status, control)
+	switch outcomeClass {
+	case lifecycleControlOutcomeClassNotFound,
+		string(factorysessionexecution.LifecycleControlOutcomeInvalidState),
+		string(factorysessionexecution.LifecycleControlOutcomeTerminalSession):
+		fs.logger.Warn("factory session lifecycle control rejected", fields...)
+	default:
+		fs.logger.Info("factory session lifecycle control", fields...)
+	}
+
+	fs.emitLiveLifecycleControlMetric(sessionID, operation, outcomeClass)
+}
+
+func lifecycleControlOutcomeClass(
+	outcome factorysessionexecution.LifecycleControlOutcome,
+	err error,
+) string {
+	if err != nil {
+		if errors.Is(err, apisurface.ErrFactorySessionNotFound) {
+			return lifecycleControlOutcomeClassNotFound
+		}
+		var controlErr *factorysessionexecution.ControlError
+		if errors.As(err, &controlErr) {
+			return string(controlErr.Outcome)
+		}
+		return "ERROR"
+	}
+	if outcome == "" {
+		return "ERROR"
+	}
+	return string(outcome)
+}
+
+const (
+	lifecycleControlOutcomeClassNotFound = "NOT_FOUND"
+)
+
+func liveLifecycleControlLogFields(
+	sessionID string,
+	operation factorysessionexecution.LifecycleControlKind,
+	outcomeClass string,
+	status factorysessionexecution.LifecycleStatus,
+	control factorysessionexecution.ControlRequest,
+) []zap.Field {
+	fields := []zap.Field{
+		zap.String("session_id", sessionID),
+		zap.String("operation", string(operation)),
+		zap.String("outcome", outcomeClass),
+	}
+	if status != "" {
+		fields = append(fields, zap.String("lifecycle_control_status", string(status)))
+	}
+	if requestID := control.RequestID; requestID != "" {
+		fields = append(fields, zap.String("request_id", requestID))
+	}
+	return fields
+}
+
+func (fs *FactoryService) emitLiveLifecycleControlMetric(
+	sessionID string,
+	operation factorysessionexecution.LifecycleControlKind,
+	outcomeClass string,
+) {
+	if fs == nil {
+		return
+	}
+	session, err := fs.requireSession(sessionID)
+	if err != nil {
+		return
+	}
+	bundle := liveSessionHandle(session).runtime
+	if bundle == nil {
+		return
+	}
+	bundle.emitMetricCounter(runtimeMetricLifecycleControl, 1, metrics.Fields{
+		Outcome: outcomeClass,
+		Reason:  string(operation),
+	})
 }
